+++ FilmSearch/app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.FilmSearch">
         <activity
-            android:name=".DetailsActivity"
+            android:name=".DetailsFragment"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
