--- FilmSearch/app/src/main/java/com/example/filmsearch/DetailsActivity.kt
@@ -1,26 +1,0 @@
-package com.example.filmsearch
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import kotlinx.android.synthetic.main.activity_details.*
-
-class DetailsActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_details)
-
-        setFilmsDetails()
-    }
-
-    private fun setFilmsDetails() {
-        //Получаем наш фильм из переданного бандла
-        val film = intent.extras?.get("film") as Film
-
-        //Устанавливаем заголовок
-        details_toolbar.title = film.title
-        //Устанавливаем картинку
-        details_poster.setImageResource(film.poster)
-        //Устанавливаем описание
-        details_description.text = film.description
-    }
-}
