+++ FilmSearch/readme.md
@@ -1,20 +1,20 @@
 ## Итоговый проект 
 
-# Модуль 24
+# Модуль 25
 
-### Задание на этот раз следующее:
+### Задание:
 
-  - Нам нужно создать `RecyclerView` для главной страницы с фильмами (пока у нас нет данных по фильмам из сети, будем использовать *mock*-данные, то есть вручную создадим нашу БД с фильмами (7 - 10 штук)).
-  - Сделаем заготовку под экран деталей фильма на основе `CoordinatorLayout`, там должен быть `Collapsing Toolbar` и «привязанные к нему» кнопки В избранное, Посмотреть позже и Поделиться.
+  - Оставить одно активити, в котором будет располагаться панель навигации и *placeholder* для фрагментов.
+  - Главный экран и экран с деталями перевести на фрагменты.
+
 
 ### Дополнительное задание ✱:
 
-  - Добавьте `DiffUtils` для нашего `RV`, на экране деталей добавьте `Snackbar` на нажатие кнопок В избранное и Посмотреть позже .
+  - Добавьте всплывающее меню о том, что пользователь сейчас покинет приложение (учитывайте стэк фрагментов).
 
 
 
 
-![Скриншот Модуль 24](./FilmSearch_24.1.png)
+![Скриншот Модуль 25](./FilmSearch_25.png)
 
-![Скриншот Модуль 24](./FilmSearch_24.2.png)
